requirements.txt
python-oauth2==1.5.211-2
